backend/pyproject.toml
[project]
name = "mandarin-pitch-detector"
version = "0.1.0"
description = ""
authors = [
    {name = "Mengyuan Xiao",email = "my333x@gmail.com"}
]
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "scipy (>=1.16.0,<2.0.0)",
    "numpy (==1.26.4)",
    "tensor (>=0.3.6,<0.4.0)",
    "crepe (>=0.0.16,<0.0.17)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "imageio (>=2.37.0,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "tensorflow (==2.16.1)",
    "praat-parselmouth (>=0.4.7,<0.5.0)",
    "fastapi (>=0.109.0,<0.110.0)",
    "uvicorn (>=0.27.0,<0.28.0)",
    "python-multipart (>=0.0.9,<0.1.0)",
]

[tool.poetry]
packages = [{include = "mandarin_pitch_detector", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
